backend/requirements.txt
django==5.0.6
djangorestframework==3.15.1
django-filter==24.2
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.4
python-slugify==8.0.4
dj-database-url==2.2.0
pytest==8.2.2
pytest-django==4.8.0
pytest-cov==5.0.0
